+++ CqlSharp/Network/BalancedConnectionStrategy.cs
@@ -13,7 +13,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using CqlSharp.Config;
 using CqlSharp.Network.Partition;
 using System.Linq;
 using System.Threading;
@@ -29,7 +28,7 @@ namespace CqlSharp.Network
     /// </summary>
     internal class BalancedConnectionStrategy : IConnectionStrategy
     {
-        private readonly ClusterConfig _config;
+        private readonly CqlConnectionStringBuilder _config;
         private readonly Ring _nodes;
         private int _connectionCount;
 
@@ -38,7 +37,7 @@ namespace CqlSharp.Network
         /// </summary>
         /// <param name="nodes"> The nodes. </param>
         /// <param name="config"> The config. </param>
-        public BalancedConnectionStrategy(Ring nodes, ClusterConfig config)
+        public BalancedConnectionStrategy(Ring nodes, CqlConnectionStringBuilder config)
         {
             _nodes = nodes;
             _config = config;